requirements.txt
playwright>=1.45
python-dotenv>=1.0
requests>=2.31

